pyproject.toml
[tool.poetry]
name = "bot"
version = "0.1.0"
description = "A simple bot made for the Créu Cat community"
authors = ["Sofia \"dzshn\" N. L. <zshn@pm.me>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^1.7.3"
psutil = "^5.8.0"
fuzzywuzzy = {extras = ["speedup"], version = "^0.18.0"}
Pillow = "^8.2.0"
numpy = "^1.20.3"

[tool.poetry.dev-dependencies]
flake8 = "^3.9.2"
yapf = "^0.31.0"
toml = "^0.10.2"

[tool.yapf]
based_on_style = "google"
column_limit = 110
blank_line_before_nested_class_or_def = false
dedent_closing_brackets = true
split_before_dot = true
coalesce_brackets = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
